% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_collect_metadata.R
\name{collect_csv_categories}
\alias{collect_csv_categories}
\title{Collect all csv file categories as metadata sheet.}
\usage{
collect_csv_categories(input_path = yspm::reference_content("data"),
  output_path = yspm::reference_content("metadata/dataset"))
}
\arguments{
\item{input_path}{A path to look for csv files. Defaults to the data
folder of the active project.}

\item{output_path}{A path to save the collected metadata to.
Defaults to the metadata folder of the active project.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
Collect all csv file categories as metadata sheet.
}
\examples{
\dontrun{
collect_csv_categories(input_path = "~/test")
}

}
