% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_file_name_constructors.R
\name{compile_data_filename}
\alias{compile_data_filename}
\title{Create a file name for data}
\usage{
compile_data_filename(name = NULL, extension = NULL)
}
\arguments{
\item{name}{A filename to create for your data. It will be forced to snake
case and the name will be augmented with the path to folder that
contains all your data.}

\item{extension}{The extension the file is using. This string will be 
forced to lower case letters only and appended to the file 
name.}
}
\value{
The function returns the constructed file path
}
\description{
The package comes with family of functions which will help you with
referencing and naming content inside of a project. This provides you 
with an unobtrusive way to keep your project clean, while adhering to 
a unified naming schema.
}
\examples{
\dontrun{
 compile_data_filename(name = "01 my first data", extension = "PNG")
}

}
