% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_naming_constructors.R
\name{compile_filename}
\alias{compile_filename}
\title{Create file names for your project}
\usage{
compile_filename(name = NULL, extension = NULL)
}
\arguments{
\item{name}{A filename to create for a plot. It will be forced to snake
case.}

\item{extension}{The extension the file is using. This string will be
forced to lower case letters only and appended to the file
name.}
}
\value{
The function returns the constructed file name
}
\description{
The package comes with family of functions which will help you with
referencing and naming content inside of a project. This provides you
with an unobtrusive way to keep your project clean, while adhering to
a unified naming schema.
}
\examples{
\dontrun{
compile_filename(name = "01 my first plot", extension = "PNG")
}

}
