% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_file_name_constructors.R
\name{compile_plot_filename}
\alias{compile_plot_filename}
\title{Create a file name for a plot}
\usage{
compile_plot_filename(name = NULL, extension = NULL)
}
\description{
Create a file name for a plot
}
