% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_default_file_contents.R
\name{construct_file_packages}
\alias{construct_file_packages}
\title{A function which fills the packages file with content}
\usage{
construct_file_packages(project_path)
}
\arguments{
\item{project_path}{The path to a project}
}
\value{
The function does not return anything; It is called for its side-
effects
}
\description{
An internally used function which fills the file in the project which contains
the packages which are required.
}
