% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_default_file_contents.R
\docType{data}
\name{content_setup_manage_projects}
\alias{content_setup_manage_projects}
\title{A function which fills the project mangement file with content}
\format{An object of class \code{character} of length 1.}
\usage{
content_setup_manage_projects
}
\description{
An internally used function which fills the file in project management file 
to get the user started after installation.
}
\keyword{datasets}
