% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_create_project.R
\name{create_project}
\alias{create_project}
\title{Create a new project}
\usage{
create_project(root_path = getwd(), project_name = NULL,
  project_path = NULL)
}
\arguments{
\item{root_path}{The path to the root folder designates the location in the
file system where the project will be stored.}

\item{project_name}{The project name is the name of the folder the predefined
structure is created in. This parameter allows you to freely name the
project. You can use it together with the compile_project_name
function to generate streamlined project names. It allows you to pass
in a constructor function. It collects all parameters names and their
values to write them as a key value pair into a metadata file called
project.dc in the metadata folder of the project. The package comes
with a default constructor called compile_project_name()}

\item{project_path}{Instead of using the root_folder and project_name arguments
the full path allows to specify the path to the folder in which which
the predefined project folder structure will be created in.}
}
\value{
This function does not return anything; It is called for its side effect
}
\description{
A function that creates a new project. A project consists of a predefined
folder structure. The structure separates into folder like data, metadata,
source and reporting to help better organize your new project.
}
\examples{
\dontrun{
create_project(
  root_path = getwd(),
  project_name = compile_project_name(
    first_name = "Max",
    last_name = "Mustermann",
    project_category = "Phd"
  )
)
}

}
