% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_inspect_the_project.R
\name{detect_data_usage}
\alias{detect_data_usage}
\title{Detect dataset usage in a project}
\usage{
detect_data_usage(data_path = yspm::project_structure("folder_data"),
  script_path = yspm::project_structure("folder_source_workflow"))
}
\arguments{
\item{data_path}{A path to the folder to search for the data in. This defaults
to the data folder of the project.}

\item{script_path}{A path to the folder to search for the scripts in. This defaults
to the workflow folder of the project.}
}
\value{
The function returns named vector pointing out dataset usage in scripts
}
\description{
It collects all names of the data files and script files in the data and scource
directory and check if there are references inthe scripts to the data. It returns
the name of the dataset file which has been used the most often. This is likely
the main data of the project. The function takes into account R scripts as well as
R-Markdown files in the workflow folder in the source directory and all files in
the data folder.
}
