% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_inspect_project.R
\name{detect_main_script}
\alias{detect_main_script}
\title{Find the main script of a project}
\usage{

  detect_main_script(path = yspm::project_structure("folder_source_workflow"))
}
\arguments{
\item{path}{A path to the folder to search for the main script in. This defaults
to the workflow folder of the project.}
}
\value{
The function returns a string with the name of the main script
}
\description{
It collects all script files in the scource directory and check if the are
referencing each other. It returns the name of the script file which has the
greates count of mentioning other scripts. This is likely the main script to be
called execute the project. The function takes into account R as well as
R-Markdown files in the workflow folder in the source directory.
}
