% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_enabled_project.R
\name{enabled_project}
\alias{enabled_project}
\title{Query information about the active project}
\usage{
enabled_project(...)
}
\arguments{
\item{...}{Pass in the names of available information pieces see
example.}
}
\description{
Query information about the active project
}
\examples{
\dontrun{
 # list all
 agwdm::enabled_project()
 # get the active project checkpoint
 agwdm::enabled_project("project_checkpoint")
}

}
