% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_helper_metadata.R
\name{get_project_metadata}
\alias{get_project_metadata}
\title{Two functions to modify metadata from a project}
\usage{
get_project_metadata(metadata, element_name)
}
\arguments{
\item{metadata}{A data frame with the project metadata read 
from the dcf file of an active project}

\item{element_name}{The name of a field in the metadata
read from the dcf file of an active project}
}
\description{
The metadata can be rad from a dcf file into a data frame. The two 
functions below can be used to query the metadata and update it if
this is required.
}
