% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_inspect_project.R
\name{inspect_project}
\alias{inspect_project}
\title{Inspect a project folder}
\usage{
inspect_project(root_path = getwd(), project_name = NULL,
  project_path = NULL)
}
\arguments{
\item{root_path}{The root folder designates the path in the filesystem
where the project is stored.}

\item{project_name}{The project name is the name of the folder the predefined
project structure has been created in. This parameter allows you to freely
name the project. You can use it together with the compile_project_name
function to get the same streamlined project names as during the project
creation.}

\item{project_path}{Instead of using the root_path and project_name arguments
the full path to the project folder can be specifified here.}
}
\description{
As enabling a project can take a while, installing all the dependencies, this
function allows to peek into a project without enabling it. It will present a
summary of packages which are required, the files which are used and metadata
as well as it can show if the project contains a library which is compatible
with the current sytem architecture.
}
\examples{
\dontrun{
inspect_project(
  root_path = getwd(),
  project_name = compile_project_name(
    first_name = "Max",
    last_name = "Mustermann",
    project_category = "Phd"
  )
)

inspect_project(
  root_path = "~/",
  project_name = "my_phd_project"
)
}

}
