% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_install_packages.R
\name{install_packages}
\alias{install_packages}
\title{A function to install packages #'}
\usage{
install_packages(name)
}
\arguments{
\item{name}{A package name or character vector with package names}
}
\value{
The function is called for its side effects
}
\description{
A function which is installing packages without triggering the reload of the
environment in R-Studio. It is a wrapper around the install.packages
function. It checks for currently loaded packages, forces them to be detached
and attaches them again after your new packages have been installed.
}
