% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_install_requirements.R
\name{install_requirements}
\alias{install_requirements}
\title{A function to install packages into a project}
\usage{
install_requirements(path = yspm::project_structure("folder_source"))
}
\arguments{
\item{path}{Provide the path folder all scripts are located in.}
}
\description{
You can call the function to install all packages that are reference in
the source code of your project. This function is also used during enabling
a project to ensure all requirements are installed. The function detects
all calls to library(), require(), import::import_from().
}
