% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_project_content.R
\name{link_content}
\alias{link_content}
\title{Link to project content like files and folders}
\usage{
link_content(path = NULL)
}
\arguments{
\item{path}{The path to the file or folder inside the project that is to be
referenced. You can start referencing for}
}
\value{
a string path pointing to the file or folder or when called
        without arguments an overview about the structure and files of the
        project
}
\description{
This function compiles paths from the root folder of the enabled project.
This allows to prevent calls to setwd() setting the working directory for 
the script and thus to create more self-contained shareable R projects.
The function does not check if the file or folder exists. You have to 
ensure this. While you notice it right away when you are using the function
to reference e.g. data to read into R you might not notice it right away
when writing.
}
\examples{
\dontrun{
link_to_content("data/01_primary")
# returns
<project_folder_name>/project/data/01_primary
}


}
