% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_install_requirements.R
\name{list_requirements}
\alias{list_requirements}
\title{A function to list packages used in a project}
\usage{
list_requirements(root_path,
  path = yspm::project_structure("folder_source"))
}
\arguments{
\item{root_path}{The path to the project.}

\item{path}{Provide the path folder all scripts are located in.}
}
\description{
You can call the function to list all packages that are reference in
the source code of your project. This function is also used during enabling
a project to ensure all requirements are installed. The function detects
all calls to library(), require(), import::from().
}
