% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_project_content.R
\name{reference_content}
\alias{reference_content}
\title{Project references}
\usage{
reference_content(path = NULL)
}
\arguments{
\item{path}{The path to the file or folder inside the project that is to be
referenced.}
}
\value{
a string path pointing to the file or folder or when called
        without arguments an overview about the structure and files of the
         project
}
\description{
This function allows to compile relative paths from the root folder of the
enabled project. This prevents setwd() calls and allows to create shareable
projects.
}
