% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_project_content.R
\name{show_content}
\alias{show_content}
\title{Project overview}
\usage{
show_content()
}
\value{
Called for its side effects. A tree with folders and files
        is shown
}
\description{
This functions provides a tree like overview about the folder structure
and files which reside in an activated project.
}
\examples{
\dontrun{
show_content()
project
data
...
...
}

}
