% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_project_code.R
\name{standardize_project_code}
\alias{standardize_project_code}
\title{Clean up the source code of your project}
\usage{

  standardize_project_code(project_path = yspm::enabled_project("project_path"))
}
\arguments{
\item{project_path}{Is the full path to the folder in which the active
project resides in.}
}
\description{
A small wrapper to the styler package which allows you easily standarize
the source code of your whole project using the tidyverse style guide
to improve code readability. The corrections include spacing, indentation,
the use of the right assignment operator... etc.
}
\examples{
\dontrun{
enable_project(...)

style_project()
}

}
