% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_style_project.R
\name{style_active_project}
\alias{style_active_project}
\title{Clean up the source code of your project}
\usage{
style_active_project()
}
\description{
A small wrapper to the nice styler package which allows you easily style
the source code of your whole project for better readability.
}
\examples{
\dontrun{
enable_project(...)

style_active_project()
}

}
