% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_use_that.R
\name{use_that}
\alias{use_that}
\title{A function that allows building modular projects}
\usage{
use_that(feature)
}
\description{
Each project is different and everybody has a different taste and needs of
packages. Thus the yspm is modular allowing to script each project to your
needs. It here uses a similar approach to the the use_this package. Calling
the function will add packages and content to your active project to enable
certain functionality.
}
