% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_handle_metadata.R
\name{write_project_metadata}
\alias{write_project_metadata}
\title{Function to rite project metadata}
\usage{
write_project_metadata(metadata, file_path)
}
\arguments{
\item{metadata}{A data frame with the project metadata read
from the dcf file of an active project}

\item{element_name}{The name of a field in the metadata
read from the dcf file of an active project}
}
\description{
A function to write a metadata dataframe into a project
metadata dcf file.
}
