% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_collect_metadata.R
\name{collect_csv_metadata}
\alias{collect_csv_metadata}
\title{Collect all csv related metadata}
\usage{
collect_csv_metadata(...)
}
\arguments{
\item{...}{Arguments passed into the functions to collect variables and
categories from csv files.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
A thin wrapper around the collector for variables and categories
}
\examples{
\dontrun{
collect_csv_metadata(...)
}

}
