% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_collect_metadata.R
\name{collect_csv_variables}
\alias{collect_csv_variables}
\title{Collect all csv variables as metadata sheet}
\usage{
collect_csv_variables(input_path = yspm::reference_content("data"),
  output_path = yspm::reference_content("metadata/dataset"))
}
\arguments{
\item{input_path}{A path to look for csv files. Defaults to the data
folder of the active project.}

\item{output_path}{A path to save the collected metadata to.
Defaults to the metadata folder of the active project.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
The function collects all csv files variable names, their class and count of mising
values. It compiles a file in the
data.
}
\examples{
\dontrun{
collect_csv_variables()
}

}
