% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_crawl_and_install_packages.R,
%   R/function_install_requirements.R
\name{install_requirements}
\alias{install_requirements}
\title{A function to crawl for required packages in a project}
\usage{
install_requirements(path = yspm::project_structure("folder_source"))

install_requirements(path = yspm::project_structure("folder_source"))
}
\arguments{
\item{path}{Provide the path folder all scripts are located in.}

\item{path}{Provide the path folder all scripts are located in.}
}
\description{
If we load a project we do not know which packages are
required to run it. This function detects all references which
are made to packages in the project. Then it checks of they are
existing in the project library. If not they are finally
installed.

If we load a project we do not know which packages are
required to run it. This function detects all references which
are made to packages in the project. Then it checks of they are
existing in the project library. If not they are finally
installed.
}
