% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_install_requirements.R
\name{install_requirements}
\alias{install_requirements}
\title{A function to install packages which are not yet installed}
\usage{
install_requirements(package_names)
}
\arguments{
\item{package_names}{a vector or list providing all package names to be installed.}
}
\description{
You can call the function to install all packages you pass in the names of
which are not yet installed in any of your active library paths. Is is used
together with list_requirements when a project is enabled to ensure that all 
the required dependencies are installed.
}
