% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_install_requirements.R
\name{list_requirements}
\alias{list_requirements}
\title{A function to list packages used in a project}
\usage{
list_requirements(search_path = getwd())
}
\arguments{
\item{search_path}{Provide the path to a folder all your scripts are located 
in. This can be e.g. the root directory of a project. The function 
searches recursively all folders for R and R-Markdown files.}
}
\description{
You can call the function to list all packages that are reference in
the source code of your project. This function is also used during enabling
a project to ensure all requirements are installed. The function detects
all calls to library(), require(), import::from().
}
