% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_lists_and_options.R
\name{project_structure}
\alias{project_structure}
\title{Query locations of folder and files}
\usage{
project_structure(...)

project_structure(...)
}
\arguments{
\item{...}{Pass in the names of available folders and files to query
for their locationl}

\item{...}{Pass in the names of available folders and files to query
for their locationl}
}
\description{
Query locations of folder and files

Query locations of folder and files
}
\examples{
\dontrun{
# list all
agwdm::project_structure()
}

\dontrun{
project_structure(...)
}

}
