% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_handle_metadata.R
\name{read_project_metadata}
\alias{read_project_metadata}
\title{Functions to read project metadata}
\usage{
read_project_metadata(file_path)
}
\arguments{
\item{file_path}{A file path to the dcf formated metadata}
}
\description{
functions to read metadata from a dcf formated file into
a data frame and write from a data frame to a new project
metadata dcf file.
}
