% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_project_content.R
\name{reference_content}
\alias{reference_content}
\title{A fuction to compile project paths}
\usage{
reference_content(path = NULL)
}
\description{
This function compiles a project path which is always constructued
from the root folder of the project. As the predefined folder structure
has the folder "project" on top this is taken as the root inside the
actual project name. This makes the project more self contained and
allows references to files which do not break.
}
