% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_setup.R
\name{setup}
\alias{setup}
\title{A function to setup a project mangement workspace}
\usage{
setup(root_path = NULL)
}
\arguments{
\item{root_path}{allows you to specify where the new project management
folder will be created.}
}
\description{
It makes sense to have all your R projects in a single location to keep a
better overview and simply their management. This function initizlizes such a
location and fils the file with example content to get you started.
You can place the project mangement in any folder you like. It makes sense
to place this e.g. somewhere in your users home directory.
}
