% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_project_content.R
\name{show_content}
\alias{show_content}
\title{Project overview}
\usage{
show_content()
}
\value{
The function is called for its side effects.
}
\description{
This functions provides a tree like representation of the folder structure and
all the files which reside in an activated projec like the linux command line
tool called tree. It helps to gain a fast overview about a project.
}
\examples{
\dontrun{
show_content()
project
data
...
...
}

}
